Course Project the road to OOAD.

+++ src/main/java/com/acme/edu/message/ByteLogMessage.java
@@ -15,7 +15,7 @@ public class ByteLogMessage extends LogMessage {
         }
     }
 
-    public long getAccumulatedValue() {
+    public int getAccumulatedValue() {
         return accumulatedValue;
     }
